move: discord 디랙토리 이동 #19 기존 recommendation -> notification 이동
