+++ src/main/java/com/project/expensemanage/notification/discord/body/embedObject/DiscordEmbedAuthor.java
@@ -1,4 +1,4 @@
-package com.project.expensemanage.notification.recommendation.discord.body.embedObject;
+package com.project.expensemanage.notification.discord.body.embedObject;
 
 import lombok.Builder;
