+++ src/main/java/com/project/expensemanage/notification/discord/body/embedObject/DiscordEmbedImage.java
@@ -1,0 +1,3 @@
+package com.project.expensemanage.notification.discord.body.embedObject;
+
+public record DiscordEmbedImage() {}
